Yet Another Big Push
Lots of changes; Updated the package.json changing the private status of the repository and the build root directory. Moved the data and images folders into the assets folder. Updated education, experience and jobs data to change the dates to strings and added a current field to all of the data to signify if the job or experience is active right now. Created a general data file with general application text data. Exported the resume into pdf files for all languages and added them to the assets files folder. Removed a bunch of branding images. Minor changes to the DataManager, including the addition of the new general.json data file. Changed some styles for the footer. Changed a slight variable misspelling in the LanguageManager. Slight update in the responsive nature of the navigation bar. Edited the languages to be two leter strings. Updated the header styles. Created a new component, ResumeDownload, which consists of download buttons for the resume PDF files. Created the ResumeDownload styles file and styled the download buttons. Changed the date font size for all items in the resume as well as updating the styles and the page to accomodate the download buttons.