Minor Changes
Swapped the Barbarian bot image to an svg. Created a new PageContainerInner item that holds and constricts the main page contents to a width of 1200px max and centers it in the screen. This applies to all the routes. Removed unused item from info.json file. Added importance attribute to some projects in projects.json that is then used in the projects page to determine the size of a project item in the grid. Added main page background for testing, might not be final. Moved all image assets contained within subfolders in the assets folder into the assets folder and changed all their routes in the code. Added a grid layout for the Portfolio page.