Optimizations and small changes
Optimized project images into a smaller, more manageable size to promote faster page loading. Edited the icon list to be displayed on the home page via the index.json file. Added a transparent image of me! This will be used later in either the about or home page. Re-introduced a box-shadow to the desktop navbar, removed the dynamic shadow addition onscroll. Added a check to the PageContainerInner to not constrain the size of the index page. Edited the website themes to be more *drastic* and more applicable to each theme name, now everything looks better and such. Currently playing around with the home page aka. index, should be done by the end of my night-shift :).