Finished 404
Finished the 404 page. All non defined routes are now contained within the 404 page. A data file for the page has been created, filled and added to the DataManager. Added a flex-grow attribute to the page containers. Added a media query to the sidebar to shrink the text size on smaller devices.