Complete
Finished the project (for now).

Fixed a minor style bug in the resume where it would not get resized with the screen size. This was caused by the previous print fix. Changed the font style on other experience descriptions to match the rest of the page. Updated portfolio styles to look better on smaller screen sizes. Added an outer Index container to contain overflow. Added a key to Courses and changed its styles to also look better on smaller screen sizes. Fixed styling for About.

Exported the resume pages as PDF in all available languages and updated the old files.